backend/requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
python-docx>=1.1.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
reportlab>=4.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.30
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
scikit-learn>=1.5.0
numpy>=1.26.0

# AI / NLP
sentence-transformers>=2.7.0
spacy>=3.7.0

# Production
slowapi>=0.1.9

# Testing
pytest>=8.0.0
httpx>=0.27.0
pytest-cov>=5.0.0
